requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
idna==3.4
pycryptodome==3.18.0
python-dotenv==1.0.0
requests==2.31.0
termcolor==2.3.0
urllib3==1.26.16
yaspin==2.5.0
paramiko~=3.3.1